requirements.txt
pytest<=6.2.3
boto3<=1.17.49
mock<=4.0.3
python-dateutil<=2.8.1
docker<=4.4.2